fix case that relies on runtime being undefined for create panel
