+++ ui/src/app/utils/runtime-presets.ts
@@ -1,4 +1,5 @@
-import {Runtime, RuntimeConfigurationType} from 'generated/fetch';
+import {Runtime, RuntimeConfigurationType, RuntimeStatus} from 'generated/fetch';
+import * as fp from 'lodash/fp';
 
 export const runtimePresets: {
   [runtimePresetName: string]: {displayName: string, runtimeTemplate: Runtime}
@@ -28,4 +29,27 @@ export const runtimePresets: {
       }
     }
   }
+};
+
+export const applyPresetOverride = (runtime) => {
+  if (!runtime) {
+    return runtime;
+  }
+
+  const newRuntime = {...runtime};
+
+  // if runtime configuration type is a default, override its config with preset values
+  // The Deleted check is so that we only update the user's runtime to the latest preset values
+  // after they delete their runtime (ex. not while its actively in use).
+  if (newRuntime && newRuntime.status === RuntimeStatus.Deleted) {
+    const runtimePresetKey = fp.keys(runtimePresets)
+    .find(key => runtimePresets[key].runtimeTemplate.configurationType === newRuntime.configurationType);
+
+    if (runtimePresetKey) {
+      newRuntime.gceConfig = runtimePresets[runtimePresetKey].runtimeTemplate.gceConfig;
+      newRuntime.dataprocConfig = runtimePresets[runtimePresetKey].runtimeTemplate.dataprocConfig;
+    }
+  }
+
+  return newRuntime;
 };
